requirements.txt
# Core dependencies for all scripts
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
jinja2>=3.1.0
pydantic>=2.0.0

# HTML parsing
lxml>=4.9.0

# Image processing (for team icons)
Pillow>=10.0.0

# AI Provider dependencies
# Install only the ones you need for your chosen provider

# For Claude (Anthropic)
anthropic>=0.39.0

# For OpenAI
openai>=1.54.0

# For Google Gemini
google-generativeai>=0.8.0
